tools/structure_vis/pyproject.toml
[project]
name = "structure-vis"
version = "0.1.0"
description = "Quick PyMOL-based visualization pipeline for PDB and mmCIF structures."
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
pymol = ["pymol-open-source; python_version < '3.13'"]

[project.scripts]
structure-vis = "structure_vis.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
